-- Fixed problem with pow function in    "src/meshes/triangle_mesh.template.cc" file that prevented    some demo drivers to compile.
git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@1078 595e1d82-adf1-4282-bac7-b6073fa2ca8e

+++ src/meshes/triangle_mesh.template.cc
@@ -2388,8 +2388,8 @@ void RefineableTriangleMesh<ELEMENT>::refine_triangulateio(
      }
     for (unsigned e=0;e<n;e++)
      {
-      unsigned level=ceil(log(target_area[e]/total_area)/log(1.0/3.0))-1;
-      double new_target_area=total_area*pow(1.0/3.0,level);
+      unsigned level=unsigned(ceil(log(target_area[e]/total_area)/log(1.0/3.0)))-1;
+      double new_target_area=total_area*pow(1.0/3.0,int(level));
       target_area[e]=new_target_area;
      }
    }
